Refromatting the code
