+++ rover_video/package.xml
@@ -15,6 +15,7 @@
   <test_depend>ament_lint_common</test_depend>
 
   <depend>rover_msgs</depend>
+  <depend>opencv2</depend>
 
   <export>
     <build_type>ament_cmake</build_type>
